system: menyelesaikan landing page

+++ app/Http/Controllers/Sigarang/Goods/CategoryController.php
@@ -124,7 +124,11 @@ class CategoryController extends BaseController
     
     public function destroy($id)
     {
+        /* @var $model Category */
         $model = Category::find($id);
+        if(count($model->goods)!=0){
+            return 'Data has been used';
+        }
         return $model->delete() ? '1' : 'Data cannot be deleted';
     }
 }
